requirements.txt

openai-whisper>=20231117
sounddevice>=0.4.6
numpy>=1.24
scipy>=1.11
pyttsx3>=2.90
openai>=1.0